Fix not being able to serve compiled svg files by Swoole...
"in case you're curious how it happened, the svg was minified by
webpack, stripping the <?xml part of the svg content. That wouldn't have
been a problem by itself, because browsers don't care, however static
files were served by a PHP script, and PHP's built in mime detecting
function detected that SVG without the xml part as just image/svg, and
so it set that as the content-type"

Fixed the issue by making sure the <?xml part of the SVG is kept.