Remove useless mcamara/laravel-localization...
It doesn't really support changing locales and saving them to cookies.
It's rather for non-SPA applications and having locales in the route
paths.
The only thing it did decently was detect the language, but it's overkill
for supporting just English and Japanese, and besides, it didn't seem to
work at all with Swoole.